requirements.txt
# Core ML/AI dependencies
transformers>=4.40.0
torch>=2.2.0
accelerate>=0.28.0
sentencepiece>=0.2.0
protobuf>=4.25.0

# Data processing
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0

# Evaluation
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
gitpython>=3.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: CUDA quantization (NOT supported on Apple Silicon/MPS)
# Uncomment if running on NVIDIA GPU:
# bitsandbytes>=0.43.0
